requirements.txt
jgcmlib>=1.0.59
tlid
requests
music21
ipython
huggingface_hub
jghfmanager>=0.1.5
tiktoken
openai
numpy
langchain
langchain_community
langchainhub
langchain-openai
arxiv
tlid
pluggy
elevenlabs
langchain-experimental
python-dotenv
llm
langsmith
pandas
wikipedia
numexpr
langgraph
IPython
numexpr
